Refactored Likes tab
