+++ apps/admin-x-activitypub/src/views/Profile/Profile.tsx
@@ -21,6 +21,7 @@ import ActivityItem from '@components/activities/ActivityItem';
 import FeedItem from '@components/feed/FeedItem';
 import FollowButton from '@components/global/FollowButton';
 import Layout from '@components/layout';
+import Likes from './components/Likes';
 import Posts from './components/Posts';
 import Separator from '@components/global/Separator';
 import ViewProfileModal from '@components/modals/ViewProfileModal';
@@ -127,6 +128,8 @@ const useInfiniteScrollTab = <TData,>({useDataHook, emptyStateLabel, emptyStateI
 };
 
 const PostsTab:React.FC<{handle?: string}> = ({handle}) => {
+    // TODO: for some reason if the last activity for the current account is a repost then it sticks on the top of the postlist
+
     // Call both hooks unconditionally
     const postsByAccountQuery = usePostsByAccount({enabled: true}).postsByAccountQuery;
     const profilePostsQuery = useProfilePostsForUser('index', handle || '');
@@ -157,78 +160,13 @@ const LikesTab: React.FC = () => {
 
     const posts = data?.pages.flatMap(page => page.posts) ?? Array.from({length: 5}, (_, index) => ({id: `placeholder-${index}`, object: {}}));
 
-    const observerRef = useRef<IntersectionObserver | null>(null);
-    const loadMoreRef = useRef<HTMLDivElement | null>(null);
-    const endLoadMoreRef = useRef<HTMLDivElement | null>(null);
-
-    // Calculate the index at which to place the loadMoreRef - This will place it ~75% through the list
-    const loadMoreIndex = Math.max(0, Math.floor(posts.length * 0.75) - 1);
-
-    useEffect(() => {
-        if (observerRef.current) {
-            observerRef.current.disconnect();
-        }
-
-        observerRef.current = new IntersectionObserver((entries) => {
-            if (entries[0].isIntersecting && hasNextPage && !isFetchingNextPage) {
-                fetchNextPage();
-            }
-        });
-
-        if (loadMoreRef.current) {
-            observerRef.current.observe(loadMoreRef.current);
-        }
-        if (endLoadMoreRef.current) {
-            observerRef.current.observe(endLoadMoreRef.current);
-        }
-
-        return () => {
-            if (observerRef.current) {
-                observerRef.current.disconnect();
-            }
-        };
-    }, [hasNextPage, isFetchingNextPage, fetchNextPage]);
-
-    return (
-        <>
-            {hasNextPage === false && posts.length === 0 && (
-                <NoValueLabel icon='heart'>
-                    You haven&apos;t liked anything yet.
-                </NoValueLabel>
-            )}
-            <ul className='mx-auto flex max-w-[640px] flex-col'>
-                {posts.map((activity, index) => (
-                    <li
-                        key={`likes-${activity.id}`}
-                        data-test-view-article
-                    >
-                        <FeedItem
-                            actor={activity.actor}
-                            allowDelete={activity.object.authored}
-                            commentCount={activity.object.replyCount}
-                            isLoading={isLoading}
-                            layout='feed'
-                            object={activity.object}
-                            repostCount={activity.object.repostCount}
-                            type={activity.type}
-                            onClick={() => handleViewContent(activity, false)}
-                            onCommentClick={() => handleViewContent(activity, true)}
-                        />
-                        {index < posts.length - 1 && <Separator />}
-                        {index === loadMoreIndex && (
-                            <div ref={loadMoreRef} className='h-1'></div>
-                        )}
-                    </li>
-                ))}
-                {isFetchingNextPage && (
-                    <li className='flex flex-col items-center justify-center space-y-4 text-center'>
-                        <LoadingIndicator size='md' />
-                    </li>
-                )}
-            </ul>
-            <div ref={endLoadMoreRef} className='h-1'></div>
-        </>
-    );
+    return <Likes
+        fetchNextPage={fetchNextPage}
+        hasNextPage={hasNextPage!}
+        isFetchingNextPage={isFetchingNextPage}
+        isLoading={isLoading}
+        posts={posts}
+    />;
 };
 
 const handleAccountClick = (handle: string) => {
@@ -351,7 +289,7 @@ const Profile: React.FC<ProfileProps> = ({}) => {
                 </div>
             )
         },
-        {
+        !params.handle && {
             id: 'likes',
             title: 'Likes',
             contents: (
